webapp/requirements.txt
azure-cosmos>=4.0.0
Flask>=1.0.2
Flask_HTTPAuth>=4.1.0
requests>=2.24.0
werkzeug>=0.15.3
